Move get_expath function from init.rs to main.rs
