+++ src/init.rs
@@ -3,15 +3,3 @@ use std::env;
 
 pub(crate) fn init() {
 }
-
-pub(crate) fn get_expath() -> PathBuf {
-    let expath = match env::var("EXPATH") {
-        Ok(path) => PathBuf::from(path),
-        Err(_) => {
-            let mut home = dirs::home_dir().expect("Home directory not found");
-            home.push(".expack");
-            home
-        },
-    };
-    expath
-}
